backend/requirements.txt
fastapi
uvicorn
aiohttp
aiofiles
pydantic
python-multipart
py7zr
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
